+++ apps/api/package.json
@@ -17,8 +17,8 @@
     "docker:build:depot": "pnpm --silent --workspace-root pnpm-context -- apps/api/Dockerfile | depot build --build-arg PACKAGE_PATH=apps/api - -t novu-api --load",
     "start": "pnpm start:dev",
     "start:dev": "nest start --watch",
-    "start:portless": "cross-env PORTLESS_INJECT_CA=1 node ../../scripts/with-portless-env.mjs pnpm exec portless run pnpm start:dev",
-    "start:portless:ngrok": "cross-env PORTLESS_INJECT_CA=1 PORTLESS_NGROK=1 node ../../scripts/with-custom-ngrok-tunnel.mjs node ../../scripts/with-portless-env.mjs pnpm exec portless run pnpm start:dev",
+    "start:portless": "cross-env PORTLESS_INJECT_CA=1 node ../../scripts/portless-dev-env.mjs pnpm exec portless run pnpm start:dev",
+    "start:portless:ngrok": "cross-env PORTLESS_INJECT_CA=1 PORTLESS_NGROK=1 node ../../scripts/portless-dev-env.mjs --manage-ngrok pnpm exec portless run pnpm start:dev",
     "start:test": "cross-env NODE_ENV=test nest start",
     "start:debug": "nest start --debug --watch --inspect",
     "start:prod": "node dist/main.js",
