+++ apps/api/src/.example.env
@@ -118,7 +118,7 @@ SCHEDULER_URL=
 SCHEDULER_API_KEY=
 SCHEDULER_CALLBACK_API_KEY=
 
-# When using `pnpm dev:portless`, scripts/with-portless-env.mjs resolves these at
+# When using `pnpm dev:portless`, scripts/portless-dev-env.mjs resolves these at
 # runtime via `portless get` (worktree-prefixed and proxy-port aware):
 #   API_ROOT_URL, FRONT_BASE_URL, DASHBOARD_URL, BETTER_AUTH_BASE_URL
 # It also sets NODE_EXTRA_CA_CERTS to ~/.portless/ca.pem so Node trusts portless HTTPS
