+++ apps/api/src/config/env.config.ts
@@ -22,13 +22,4 @@ if (dotenvPath) {
   dotenv.config({ path: dotenvPath });
 }
 
-// Portless 0.14+ injects PORTLESS_NGROK_URL when started with --ngrok, and
-// with-portless-env sets it for reserved domains (PORTLESS_NGROK_DOMAIN).
-// Bridge to AGENT_API_HOSTNAME so agent webhooks/OAuth use the public tunnel URL.
-const portlessNgrokUrl = process.env.PORTLESS_NGROK_URL?.trim();
-
-if (!process.env.AGENT_API_HOSTNAME?.trim() && portlessNgrokUrl) {
-  process.env.AGENT_API_HOSTNAME = portlessNgrokUrl;
-}
-
 export const CONTEXT_PATH = getContextPath(NovuComponentEnum.API);
