+++ apps/dashboard/.example.env
@@ -6,7 +6,7 @@ VITE_LAUNCH_DARKLY_CLIENT_SIDE_ID=
 VITE_HUBSPOT_EMBED=
 VITE_API_HOSTNAME=http://localhost:3000
 # Public API URL for agent webhooks/OAuth. Set manually, or use `pnpm dev:portless:ngrok`
-# to have scripts/with-portless-env.mjs inject it from portless automatically.
+# to have scripts/portless-dev-env.mjs inject it from portless automatically.
 # For a reserved ngrok domain: PORTLESS_NGROK_DOMAIN=my-api.ngrok.app pnpm dev:portless:ngrok
 # VITE_AGENT_API_HOSTNAME=https://your-tunnel.ngrok.app
 VITE_WEBSOCKET_HOSTNAME=http://localhost:3002
@@ -50,7 +50,7 @@ VITE_REGION_FLAG_SG=🇸🇬
 # VITE_REGION_FLAG_<CODE>=<emoji-flag>
 # See MULTI_REGION_SETUP.md for detailed instructions
 
-# When using `pnpm dev:portless`, scripts/with-portless-env.mjs resolves these at
+# When using `pnpm dev:portless`, scripts/portless-dev-env.mjs resolves these at
 # runtime via `portless get` (worktree-prefixed and proxy-port aware):
 #   VITE_API_HOSTNAME, VITE_WEBSOCKET_HOSTNAME, VITE_DASHBOARD_URL, VITE_BETTER_AUTH_BASE_URL
 #
