+++ apps/worker/package.json
@@ -11,7 +11,7 @@
     "docker:build": "pnpm --silent --workspace-root pnpm-context -- apps/worker/Dockerfile | BULL_MQ_PRO_NPM_TOKEN=${BULL_MQ_PRO_NPM_TOKEN} docker buildx build --secret id=BULL_MQ_PRO_NPM_TOKEN --build-arg PACKAGE_PATH=apps/worker - -t novu-worker --load $DOCKER_BUILD_ARGUMENTS",
     "docker:build:depot": "pnpm --silent --workspace-root pnpm-context -- apps/worker/Dockerfile | depot build --build-arg PACKAGE_PATH=apps/worker - -t novu-worker --load",
     "start": "pnpm start:dev",
-    "start:portless": "cross-env PORTLESS_INJECT_CA=1 node ../../scripts/with-portless-env.mjs pnpm start:dev",
+    "start:portless": "cross-env PORTLESS_INJECT_CA=1 node ../../scripts/portless-dev-env.mjs pnpm start:dev",
     "start:dev": "nest start --watch",
     "start:test": "cross-env NODE_ENV=test nest start",
     "start:debug": "nest start --debug --watch",
