+++ apps/worker/src/.example.env
@@ -80,7 +80,7 @@ API_ROOT_URL=http://localhost:3000
 # Bridge execution resolves the API base from API_INTERNAL_ORIGIN, then API_ROOT_URL.
 #
 # When using `pnpm dev:portless`, start the worker via `start:portless` so
-# scripts/with-portless-env.mjs injects API_ROOT_URL, API_INTERNAL_ORIGIN, and
+# scripts/portless-dev-env.mjs injects API_ROOT_URL, API_INTERNAL_ORIGIN, and
 # related URLs from `portless get` (same as the API service).
 #
 # With `pnpm dev:portless:ngrok`, AGENT_API_HOSTNAME is also injected from the
