+++ mprocs.yaml
@@ -1,6 +1,6 @@
 procs:
   api:
-    shell: pnpm --filter @novu/api-service start:portless
+    shell: pnpm --filter @novu/api-service ${API_PORTLESS_SCRIPT:-start:portless}
   dashboard:
     shell: pnpm --filter @novu/dashboard start:portless
   worker:
