+++ package.json
@@ -20,7 +20,7 @@
     "commit": "cz",
     "dev-environment-setup": "./scripts/dev-environment-setup.sh",
     "dev:portless": "mprocs",
-    "dev:portless:ngrok": "cross-env PORTLESS_EXPECT_NGROK=1 mprocs -c mprocs-ngrok.yaml",
+    "dev:portless:ngrok": "cross-env PORTLESS_NGROK=1 API_PORTLESS_SCRIPT=start:portless:ngrok mprocs",
     "docker:build": "pnpm -r --if-present --parallel docker:build",
     "g:module": "hygen module new --name=$pnpm_config_name",
     "g:usecase": "hygen usecase new --name=$pnpm_config_name --module=$pnpm_config_module",
@@ -45,7 +45,7 @@
     "packages:set-workspace-protocol": "node scripts/set-package-dependencies.mjs workspace:*",
     "pnpm-context": "node scripts/pnpm-context.mjs",
     "portless:setup": "pnpm exec portless trust && pnpm exec portless service install",
-    "portless:ngrok:url": "node scripts/portless-ngrok-url.mjs api.novu",
+    "portless:ngrok:url": "node scripts/portless-ngrok.mjs url api.novu",
     "prebuild": "nx run-many --target=prebuild --all",
     "preinstall": "npx only-allow pnpm",
     "preview:pkg:build": "IS_PREVIEW=true nx affected -t build --base=origin/next --head=HEAD --exclude='*,!tag:type:package'",
